requirements.txt
bson==0.5.10
pymongo==3.12.3
python-dotenv==0.20.0
jieba==0.42.1
scikit-learn==1.1.1